Math Fix for Velocity calc in Component.cs
I don't believe this component is currently calculating our total speed correctly, because it's doing the Sqrt with all values together instead of separately. For example, I think your Math.Sqrt(x * x + z * z) should instead be Math.Sqrt(x * x) + Math.Sqrt(z * z), because, let's assume X=5 Y=10 for a speed of 15. If we do your formula √(5²+10²) we actually get 11 instead of 15, but if we do the Sqrts separately like this: √(5²)+√(10²) we get the 15 we want. I tried to adjust all of your formulas in this file for this change, but I didn't actually compile to test them. I think they're correct, but I can't be 100% sure right now.